refactor(workbench): extract Settings overlay into WorkbenchSettingsStage
Move the SettingsPage mount and prop surface out of AppWorkbench so settings chrome can change without opening the 17k-line host. Settings state and navigation stay on the host.
